+++ environments/gloo-gateway/gateway-api/1.17/additional-cluster-1/gloo-agent/test.sh
@@ -4,7 +4,7 @@
 if [[ ${gloo_mesh_version} == "" ]]
   then
     # provide gloo_mesh_version variable
-    echo "Please provide the gloo_mesh_version to use (i.e. v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62):"
+    echo "Please provide the gloo_mesh_version to use (i.e. v2.7.0-rc2):"
     read gloo_mesh_version
 fi
 
@@ -104,8 +104,8 @@ spec:
         glooAnalyzer:
             enabled: true
 
-    repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62
+    repoURL: https://storage.googleapis.com/gloo-platform/helm-charts
+    targetRevision: v2.7.0-rc2
   syncPolicy:
     automated:
       prune: true
@@ -162,8 +162,8 @@ spec:
         telemetryCollectorCustomization:
             skipVerify: true
 
-    repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62
+    repoURL: https://storage.googleapis.com/gloo-platform/helm-charts
+    targetRevision: v2.7.0-rc2
   syncPolicy:
     automated:
       prune: true