+++ environments/gloo-platform/gwapi-mgmt-gm-workers/cluster2/gloo-platform/init.sh
@@ -7,7 +7,7 @@ echo "deploy and register gloo-mesh agent and addons"
 if [[ ${gloo_mesh_version} == "" ]]
   then
     # provide gloo_mesh_version variable
-    echo "Please provide the gloo_mesh_version to use (i.e. v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62):"
+    echo "Please provide the gloo_mesh_version to use (i.e. v2.7.0-rc2):"
     read gloo_mesh_version
 fi
 
@@ -74,8 +74,8 @@ spec:
                   name: null
                   namespace: null
                   
-    repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62
+    repoURL: https://storage.googleapis.com/gloo-platform/helm-charts
+    targetRevision: v2.7.0-rc2
   syncPolicy:
     automated:
       prune: true
@@ -123,7 +123,7 @@ spec:
             image:
               pullPolicy: IfNotPresent
               repository: gcr.io/gloo-mesh/gloo-otel-collector
-              tag: v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62
+              tag: v2.7.0-rc2
             config:
                 exporters:
                     otlp:
@@ -144,8 +144,8 @@ spec:
                       type: DirectoryOrCreate
                   name: cilium-run
                   
-    repoURL: https://storage.googleapis.com/gloo-platform-dev/platform-charts/helm-charts
-    targetRevision: v2.8.0-beta0-2025-02-14-19265-new-cluster-details-0966fa9b62
+    repoURL: https://storage.googleapis.com/gloo-platform/helm-charts
+    targetRevision: v2.7.0-rc2
   syncPolicy:
     automated:
       prune: true